+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+##
+
+* the regular expression of the tokenizer is now configurable via config.json
+
 ## 1.10 strong negation - 01-06-2020
 
 * strong negation: the negation operator for relations